Conflicts:
	doc/odt/OpenCPI_Application_Development_Guide.fodt

+++ build/prerequisites/lzma/install-lzma.sh
@@ -1,4 +1,4 @@
-#!/bin/bash --noprofile
+#!/bin/bash
 # This file is protected by Copyright. Please refer to the COPYRIGHT file
 # distributed with this source distribution.
 #
@@ -17,29 +17,41 @@
 # You should have received a copy of the GNU Lesser General Public License along
 # with this program. If not, see <http://www.gnu.org/licenses/>.
 
-lzma_version=5.2.3
-dir=xz-$lzma_version
-[ -z "$OCPI_CDK_DIR" ] && echo Environment variable OCPI_CDK_DIR not set && exit 1
-source $OCPI_CDK_DIR/scripts/setup-prerequisite.sh \
-       "$1" \
-       lzma \
-       "LZMA compression library" \
-       https://tukaani.org/xz \
-       $dir.tar.gz \
-       $dir \
-       1
-../configure  ${OcpiCrossHost:+--host=$OcpiCrossHost} \
-  --prefix=$OcpiInstallDir --exec-prefix=$OcpiInstallExecDir \
+[ -z "$OCPI_CDK_DIR" ] && echo 'Environment variable OCPI_CDK_DIR not set' && exit 1
+
+target_platform="$1"
+name=lzma
+version=5.2.5  # latest as of 09/02/2020
+pkg_name="xz-$version"
+description='LZMA2 compression library'
+dl_url="https://opencpi-repo.s3.us-east-2.amazonaws.com/prerequisites/${pkg_name}.tar.gz"
+extracted_dir="$pkg_name"
+cross_build=1
+
+# Download and extract source
+source "$OCPI_CDK_DIR/scripts/setup-prerequisite.sh" \
+       "$target_platform" \
+       "$name" \
+       "$description" \
+       "${dl_url%/*}" \
+       "${dl_url##*/}" \
+       "$extracted_dir" \
+       "$cross_build"
+
+# Configure/Make/Install
+../configure "${OcpiCrossHost:+--host=$OcpiCrossHost}" \
+  --prefix="$OcpiInstallDir" --exec-prefix="$OcpiInstallExecDir" \
   --enable-shared=yes --enable-static --disable-symbol-versions \
   --disable-xz --disable-xzdec --disable-lzmadec --disable-lzmainfo --disable-lzma-links \
   --disable-scripts --disable-doc \
   --with-pic=liblzma \
   CFLAGS="-g -fPIC" CXXFLAGS="-g -fPIC" # why doesn't with-pic to this?
-# Leaving -j unconstrained goes very badly on some systems, including MacOS
 make -j4
 make install
+
+# Cleanup
 # lzma creates an empty directory even when we have disabled the executables
-rm -r -f $OcpiInstallExecDir/bin
-# this is not needed in our sandbox
-rm -r -f $OcpiInstallExecDir/lib/pkgconfig
-rm -f $OcpiInstallExecDir/lib/*.la
+rm -rf "${OcpiInstallExecDir:?}/bin"
+rm -rf "${OcpiInstallExecDir:?}/lib/pkgconfig"
+rm  -f "${OcpiInstallExecDir:?}/lib/*.la"
+rm -rf "${OcpiInstallDir:?}/share"
